Server/Spells, Talents, Chat, Emotes: Refactored Chat system. Refactored Emotes system. Implemented some opcodes for Mail, Emotes, Duels. Implemented new Spell power cost and difficulty systems, a lot still to do. Please note Spell and Talent systems are WIP heavily, so build is not possible yet!

+++ src/server/game/Entities/GameObject/GameObject.cpp
@@ -1695,7 +1695,7 @@ void GameObject::CastSpell(Unit* target, uint32 spellId)
         return;
 
     bool self = false;
-    for (uint8 i = 0; i < MAX_SPELL_EFFECTS; ++i)
+    for (uint32 i = 0; i < MAX_SPELL_EFFECTS; ++i)
     {
         if (spellInfo->Effects[i].TargetA.GetTarget() == TARGET_UNIT_CASTER)
         {
